# Micro optimizations in ATen