* Compare typeid instead of using dynamic_cast
* Mark derived TensorImpl classes as final
* Use tensor->nDimension instead of THTensor_(nDimension)

For a 1x1 tensor:
- Shaves off ~30 ns (out of 190 ns) for `addmm_`
- Shaves off ~60 ns (out of 625 ns) for `addmm`